spikes/p0-006-gemma4-evaluation/requirements.txt
# P0-006: Gemma 4 Model Evaluation Spike
# Python 3.10+ required

# MediaPipe GenAI for on-device inference (Python side for embedding parity)
mediapipe>=0.10.21

# Google AI for Gemma model access (download only)
google-generativeai>=0.8.0

# Embedding and retrieval
numpy>=1.26.0
faiss-cpu>=1.8.0
scipy>=1.12.0

# Model downloading
huggingface-hub>=0.25.0

# Memory and system monitoring
psutil>=6.0.0

# Configuration
pyyaml>=6.0

# Output formatting
tabulate>=0.9.0
jinja2>=3.1.0

# Testing / validation
scikit-learn>=1.4.0
